Update tts_win.py
There's an issue that in reading Chinese language, I mean in UTF-8 format, it is pronouncing completely wrong characters.
This is because the Chinese charactors are only shown in UTF-8 code in sublime.
Before calling the sentence.encode() method, it is already utf-8 in sentence variable. It is encoded again when use the encode(), wrong charactors are recoganized by SAPI.
To fix it, I just replace the "sentence.encode('utf-8')" with "sentence" in this file.